Add extra request headers
This includes:
  X-MWU-Mobilize
  X-Forwarded-For

These are added to the http request headers before passing along to the source
website, with appropriate values.  Also this commits sets up a structure that
makes it pretty easy to add new ones in the future.

This augments what was already in place with mobilize.headers.request, because
what existed before could only modify existing request headers; now we can add
new ones.