Moplate response body: convert type from django SafeUnicode to bytes
Converts the type of final_body from type django.utils.safestring.SafeUnicode
to network-friendly bytes.

This is only needed because we're using Django templates for text rendering.
Someday if/when this is no longer the case, this conversion can be removed for
what will probably be a significant performance improvement.